Update to apache2 ~> 8.0 cookbook
This brings us up to date with the latest apache2 cookbook which
included a major refactor in 6.0.0 removing all of the definitions and
recipe with proper resources. Instead of using the apache2_default_site
resource, directly use a template and then enable the config file using
the apache2_site resource. This gives us the most flexibility.

Additional fixes:
- Install mod_wsgi as a package on RHEL since there is no built-in
  resource for it.
- Convert web_app to template and subscribe to restarting apache.
- Remove erroneous restart for ironic-api service on packages
- Properly set service password and auth URL
- Improve tests for ironic.conf
- Add missing apache2 depend
- Add missing api RHEL ChefSpec tests

Change-Id: I198e2c211630e190bf2a992b3dc6b6c5afaf54e8